tensormap-backend/pyproject.toml
[project]
name = "tensormap-backend"
version = "0.1.0"
description = "TensorMap Python Server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlmodel>=0.0.22",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    "pydantic-settings>=2.0",
    "python-socketio>=5.10",
    "numpy>=1.24",
    "pandas>=2.1",
    "tensorflow>=2.16",
    "pyyaml>=6.0",
    "flatten-json>=0.1.13",
    "Jinja2>=3.1",
    "python-multipart>=0.0.9",
    "Werkzeug>=3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pre-commit>=4.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
